requirements.txt
AWSIoTPythonSDK==1.4.7
pyusb==1.0.2
RPi.GPIO==0.7.0

